requirements.txt
alembic==1.9.1
attrs==22.2.0
autopep8==2.0.1
blinker==1.5
certifi==2022.12.7
chardet==5.1.0
charset-normalizer==2.1.1
click==8.1.3
coverage==7.0.1
exceptiongroup==1.1.0
Flask==2.2.2
Flask-Cors==3.0.10
Flask-Migrate==4.0.0
Flask-SQLAlchemy==3.0.2
idna==3.4
iniconfig==1.1.1
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.1
packaging==22.0
pipi==1.0.1
pluggy==1.0.0
psycopg2-binary==2.9.5
pycodestyle==2.10.0
pytest==7.2.0
pytest-cov==4.0.0
python-dateutil==2.8.2
python-dotenv==0.21.0
python-editor==1.0.4
requests==2.28.1
six==1.16.0
SQLAlchemy==1.4.45
tabulate==0.9.0
toml==0.10.2
tomli==2.0.1
urllib3==1.26.13
Werkzeug==2.2.2
